requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic[email]==2.9.0
pydantic-settings==2.11.0
python-dotenv==1.0.1
supabase==2.9.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
